jest, jest-babel and jest html.reportes added.
