+++ docs/MiniMemory_MiniMemory.html
@@ -3,7 +3,7 @@
 <head>
     
     <meta charset="utf-8">
-    <title>lib/fullscreen.js - Documentation</title>
+    <title>MiniMemory - Documentation</title>
     
     
     <script src="scripts/prettify/prettify.js"></script>
@@ -33,71 +33,140 @@
 
 <div id="main">
     
-    <h1 class="page-title">lib/fullscreen.js</h1>
+    <h1 class="page-title">MiniMemory</h1>
     
 
     
 
 
 
-    
-    <section>
-        <article>
-            <pre class="prettyprint source linenums"><code>/**
- * @classdesc Cross-Browser available fullscreen support.
- * @class
- */
-export class FullScreen {
-  /**
-   * @summary Create cross-fulscreen-utility.
-   */
-  constructor() {
-    const doc = window.document;
-    const docEl = doc.documentElement;
-    this.doc = doc;
-    this.docEl = docEl;
-    this.requestFullScreen =
-      docEl.requestFullscreen ||
-      docEl.mozRequestFullScreen ||
-      docEl.webkitRequestFullScreen ||
-      docEl.msRequestFullscreen;
-    this.cancelFullScreen =
-      doc.exitFullscreen ||
-      doc.mozCancelFullScreen ||
-      doc.webkitExitFullscreen ||
-      doc.msExitFullscreen;
-  }
-  /**
-   * @summary Enters to fulscreen mode.
-   */
-  enter() {
-    if (
-      !this.doc.fullscreenElement &amp;&amp;
-      !this.doc.mozFullScreenElement &amp;&amp;
-      !this.doc.webkitFullscreenElement &amp;&amp;
-      !this.doc.msFullscreenElement
-    ) {
-      this.requestFullScreen.call(this.docEl);
-    }
-  }
-  /**
-   * @summary Exits from fulscreen mode.
-   */
-  exit() {
-    if (
-      this.doc.fullscreenElement ||
-      this.doc.mozFullScreenElement ||
-      this.doc.webkitFullscreenElement ||
-      this.doc.msFullscreenElement
-    ) {
-      this.cancelFullScreen.call(this.doc);
-    }
-  }
-}
 
-</code></pre>
-        </article>
-    </section>
+<section>
+
+<header>
+    
+        <h2>
+            <span class="ancestors"><a href="MiniMemory.html">MiniMemory</a>#</span>
+        
+        MiniMemory
+        </h2>
+        
+    
+</header>
+
+<article>
+    
+        <div class="container-overview">
+        
+            
+
+    
+
+    <h4 class="name" id="MiniMemory"><span class="type-signature"></span>new MiniMemory<span class="signature">()</span><span class="type-signature"></span></h4>
+
+    
+    <p class="summary"><p>Contruct game</p>.</p>
+    
+
+
+
+
+<dl class="details">
+
+    
+    <dt class="tag-source">Source:</dt>
+    <dd class="tag-source"><ul class="dummy"><li>
+        <a href="mini-memory.js.html">mini-memory.js</a>, <a href="mini-memory.js.html#line33">line 33</a>
+    </li></ul></dd>
+    
+
+    
+
+    
+
+    
+
+    
+
+    
+
+    
+
+    
+
+    
+
+    
+
+    
+
+    
+
+    
+
+    
+
+    
+
+    
+</dl>
+
+
+
+
+
+<div class="description">
+    <p>Contruct game</p>
+</div>
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+        
+        </div>
+    
+
+    
+
+    
+
+    
+
+     
+
+    
+
+    
+
+    
+
+    
+
+    
+</article>
+
+</section>
