+++ package.json
@@ -7,18 +7,20 @@
   "license": "MIT",
   "private": false,
   "scripts": {
-    "dev": "webpack-dev-server --mode development",
+    "dev": "webpack-dev-server --mode development && yarn test",
     "build": "rimraf ./public && webpack -c \"webpack.config.js\" --mode production",
     "lint": "eslint ./src",
     "serve": "browser-sync start --server \"./public\"",
     "serve-docs": "yarn doc && browser-sync start --server \"./docs\"",
-    "doc": "rimraf ./docs && jsdoc -c \"./jsdoc.config.json\""
+    "doc": "rimraf ./docs && jsdoc -c \"./jsdoc.config.json\"",
+    "test": "jest"
   },
   "devDependencies": {
     "@babel/core": "^7.2.2",
     "@babel/preset-env": "^7.2.3",
     "@pixi/jsdoc-template": "^2.4.2",
     "autoprefixer": "^9.4.4",
+    "babel-jest": "^24.5.0",
     "babel-loader": "^8.0.0-beta",
     "css-loader": "^2.1.0",
     "docdash": "^1.0.2",
@@ -26,8 +28,11 @@
     "eslint-config-google": "^0.11.0",
     "file-loader": "^3.0.1",
     "html-loader": "^0.5.5",
+    "html-loader-jest": "^0.2.1",
     "html-webpack-plugin": "^3.2.0",
     "husky": "^1.3.1",
+    "jest": "^24.5.0",
+    "jest-html-reporter": "^2.5.0",
     "jsdoc": "^3.5.5",
     "nodemon": "^1.18.9",
     "postcss-loader": "^3.0.0",
@@ -41,7 +46,7 @@
   },
   "husky": {
     "hooks": {
-      "pre-commit": "yarn doc && yarn lint --fix && yarn build && git add ."
+      "pre-commit": "yarn test && yarn doc && yarn lint --fix && yarn build && git add ."
     }
   },
   "repository": "https://github.com/softberry/memory-game.git",
