+++ public/index.html
@@ -20,5 +20,5 @@
   </head>
   <body>
     <mini-memory matrix="2x2" settings="yes"> </mini-memory>
-  <script type="text/javascript" src="/mini-memory.min.js?b50fb8bac4b3b9542746"></script></body>
+  <script type="text/javascript" src="/mini-memory.min.js?bdb65f971a418c25c650"></script></body>
 </html>
