+++ src/test/lib/__mocks__/canvas-list.js
@@ -1,0 +1,9 @@
+module.exports = [
+  { offsetWidth: '10px', offsetHeight: '40' },
+  { offsetWidth: '20px', offsetHeight: '90' },
+  { offsetWidth: '30px', offsetHeight: '99' },
+  { offsetWidth: '50px', offsetHeight: 50 },
+  { offsetWidth: '70px', offsetHeight: 300 },
+  { offsetWidth: '90px', offsetHeight: 500 },
+  { offsetWidth: '120px', offsetHeight: 200 },
+];
